Added more methods to parent controller Also removed definitions for db in each controller.
