+++ SeaSharpe_CVGS/SeaSharpe_CVGS/Controllers/Controller.cs
@@ -1,26 +1,94 @@
-﻿using SeaSharpe_CVGS.Models;
-using System;
-using System.Collections.Generic;
+﻿using System;
+using System.Globalization;
 using System.Linq;
+using System.Security.Claims;
+using System.Threading.Tasks;
 using System.Web;
+using System.Web.Mvc;
+using Microsoft.AspNet.Identity;
+using Microsoft.AspNet.Identity.Owin;
+using Microsoft.Owin.Security;
+using SeaSharpe_CVGS.Models;
 
 namespace SeaSharpe_CVGS.Controllers
 {
     public class Controller : System.Web.Mvc.Controller
     {
         ApplicationDbContext _db;
+        ApplicationSignInManager _signInManager;
+        ApplicationUserManager _userManager;
 
+        /// <summary>
+        /// For backwards compatibility
+        /// </summary>
         public ApplicationDbContext db
+        {
+            get
+            {
+                return _db ?? (_db = new ApplicationDbContext());
+            }
+
+            protected set
+            {
+                _db = value;
+            }
+        }
+
+        public ApplicationDbContext DbContext
         {
             get
             {
                 return _db ?? new ApplicationDbContext();
             }
 
-            set
+            protected set
             {
                 _db = value;
             }
+        }
+
+        public ApplicationSignInManager SignInManager
+        {
+            get
+            {
+                return _signInManager ?? HttpContext.GetOwinContext().Get<ApplicationSignInManager>();
+            }
+            protected set
+            {
+                _signInManager = value;
+            }
+        }
+
+        public ApplicationUserManager UserManager
+        {
+            get
+            {
+                return _userManager ?? HttpContext.GetOwinContext().GetUserManager<ApplicationUserManager>();
+            }
+            protected set
+            {
+                _userManager = value;
+            }
+        }
+
+        protected override void Dispose(bool disposing)
+        {
+            if (disposing)
+            {
+                if (_userManager != null)
+                {
+                    _userManager.Dispose();
+                    _userManager = null;
+                }
+
+                if (_signInManager != null)
+                {
+                    _signInManager.Dispose();
+                    _signInManager = null;
+                }
+            }
+
+            base.Dispose(disposing);
         }
     }
 }
