+++ SeaSharpe_CVGS/SeaSharpe_CVGS/Controllers/EventController.cs
@@ -12,8 +12,6 @@ namespace SeaSharpe_CVGS.Controllers
 {
     public class EventController : Controller
     {
-        private ApplicationDbContext db = new ApplicationDbContext();
-
         #region Multiple Roles
         /// <summary>
         /// checks authorization and redirects to appropriate page
