+++ SeaSharpe_CVGS/SeaSharpe_CVGS/Controllers/OrderController.cs
@@ -12,8 +12,6 @@ namespace SeaSharpe_CVGS.Controllers
 {
     public class OrderController : Controller
     {
-        private ApplicationDbContext db = new ApplicationDbContext();
-
         #region Multiple Roles
         /// <summary>
         /// checks authorization and redirects to appropriate page
